+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Improvements
 * Using ruff to enforce existence of public classes' docstrings [PR #1034](https://github.com/catalystneuro/neuroconv/pull/1034)
 * Separated tests that use external data by modality [PR #1049](https://github.com/catalystneuro/neuroconv/pull/1049)
-
+* Add writing to zarr test for to the data tests [PR #1056](https://github.com/catalystneuro/neuroconv/pull/1056)
 
 
 ## v0.6.1 (August 30, 2024)